[line-clamp] Make `ComputedStyle::LineClamp` ignore invalid `-webkit-line-clamp`
When adding the `line-clamp` property in crrev.com/c/5317993, the
existing `ComputedStyle::LineClamp` method was changed to return the
value of this property, if not `none`, or otherwise the value of
`-webkit-line-clamp`. And similarly for `ComputedStyle::HasLineClamp`.

However, while the `line-clamp` property applies to every block
container, the `-webkit-line-clamp` property only takes effect for
boxes with the `display: -webkit-box` and
`-webkit-box-orient: vertical` styles. That meant that in practice,
uses of `ComputedStyle::LineClamp` had to be guarded behind conditions
that checked this.

This patch changes those two methods to only use the value of
`-webkit-line-clamp` when those conditions hold, returning 0
(corresponding to `none`) otherwise. The raw values of the properties
can be obtained with `StandardLineClamp` (for the unprefixed property)
and `WebkitLineClamp` if needed.

Bug: 40336192
Change-Id: Iabc08cd4bcb114dbfeb566f15966c2f7565ee6d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5457331
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Andreu Botella <abotella@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1287964}
NOKEYCHECK=True
GitOrigin-RevId: 0a0454930d8635575d34773b9ea9503e6d0d67d4