[A11y] Only top down creation
Only allow parents to create and add AXObjects in the tree, when the parent computes its children via UpdateChildrenIfNecessary(). The only
exception is the root, which is created at the first opportunity in
clean layout processing.

In combination with earlier CLs, holes can no longer
exist in the tree structure, e.g. the parent chain for a node. This
in turn allows removal of RepairChildrenOfIncludedChildren() as the
tree will never need repair.

This fixes a number of crashes where we tried to serialize nodes
that weren't included in the tree (only ever building the tree
top down leads to consistency on all nodes being included).

Fixed: 327486173,332369109,331366002,332635500
Change-Id: I858bb45c409414f451574c01e366219fe46bfc03
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5319296
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288020}
NOKEYCHECK=True
GitOrigin-RevId: 6bed1d09e35563e8d2b4f7efe89795d81ec9e4f5