Reland: Roll nearby-connections from 33e682f24427 to f788b891908d (52 revisions)
This is a reland of crrev.com/c/5434516. It was reverted due to ASan
failures. This is fixed by not casting a 16-byte array to char*, causing
C++ to think a NUL terminator existed at a non-existent 17th index.
This is fixed by instead it is casted to std::string.

Patch 1 is the original CL. Patch 2 is the fix.

https://chromium.googlesource.com/external/github.com/google/nearby-connections.git/+log/33e682f24427..f788b891908d

The latest nearby uprev deprecates ::nearby::internal::Metadata
and replaces it with ::nearby::internal::DeviceIdentityMetaData.

This CL updates chromium to use this new proto message, and transfer
it across the nearby_presence.mojom pipe.

This is primarily a mechanical change, except for
LocalDeviceDataProviderImpl. This class's device ID generation has
been adjusted to match the expected format of the device_id field
on DeviceIdentityMetaData. This does change the type of device id
we're uploading to the Nearby Presence server, but that is about
to deprecated and moved as well -- and we're not breaking anything
that is being used in production.

This CL also makes makes minor changes to
//third_party/nearby/BUILD.gn to account for a recent file refactor.

Bug: 311430390
Test: unit tests, asan test, manual testing on DUT.
Change-Id: Ia8531c71cfb5cad0a53f347bcadceda1f60494b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5449851
Reviewed-by: Juliet Lévesque <julietlevesque@google.com>
Commit-Queue: Ryan Hansberry <hansberry@chromium.org>
Reviewed-by: Brendon Tiszka <tiszka@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288180}
NOKEYCHECK=True
GitOrigin-RevId: 99706a27d063e116b965c056ef8236d204e1161b