Re-Enable AutoSizeLazyLoadedImages
This patch enables the AutoSizeLazyLoadedImages flag by default.

When enabled, lazy loaded images with `sizes=auto` will use the layout
width of the image to select the source url from the srcset. For this
use case the image's width and height should be provided.

This was initially enabled in:
https://chromium-review.googlesource.com/c/chromium/src/+/5100579

And switched back to experimental in:
https://chromium-review.googlesource.com/c/chromium/src/+/5241231
due to several image skewing bugs reported.

While investigating mitigation options for the skewing issues, it was
ultimately discovered that they were caused by the interactions between
contain-intrinsic-size and object-fit. This was brought up to the
CSS WG in issue:
https://github.com/w3c/csswg-drafts/issues/10116
The resolution was that contain-intrinsic-size should not affect
object-fit.

This was implemented in:
https://chromium-review.googlesource.com/c/chromium/src/+/5442533
guarded by flag NoIntrinsicSizeOverride, which
AutoSizeLazyLoadedImages depends on.

Spec:
https://html.spec.whatwg.org/#sizes-attributes

Chrome Status:
https://chromestatus.com/feature/5191555708616704

Intent to Ship:
https://groups.google.com/a/chromium.org/g/blink-dev/c/OAsmCbjPJz0/m/jzuTJzs1AAAJ

R=​pdr

Bug: 1359051
Change-Id: I9463ec0f6ee40809dad9e7b359c2ba05bea1a2d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5458907
Commit-Queue: Traian Captan <tcaptan@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288450}
NOKEYCHECK=True
GitOrigin-RevId: 359c83b5cfdd20d27c938c279b298d83904e3e5b