[URL] Rewrite base-url-javascript.html and add variants
http/tests/security/base-url-javascript.html is one of the tests which
would fail if kStandardCompliandNonSpecialSchemeURLParser flag is
enabled [^1] [^2].

This test is using javascript URL, "javascript:// This is JavaScript",
as a base URL. Currently, this is a valid JavaScript URL, however,
once the StandardCompliandNonSpecialSchemeURLParser feature is
enabled, this URL becomes an invalid URL because its host part
contains a space character, which is not allowed. That's the reason
this test would fail if the flag is enabled.

For the comparison purposes, this CL rewrote and separated this test
into three tests, as follows:

1. base-url-https-invalid.html: Using an invalid https URL,
   "https://[xxx]".
2. base-url-javascript-invalid.html: Using an invalid javascript URL,
   "javascript:// (space) " (hierarchical non-special URL).
3. base-url-javascript-valid.html: Using a valid javascript URL,
   "javascript:..." (opaque path URL).

Given that the current behaviors of Chromium is:

- 1. If a base URL is invalid (in the sense of the URL Standard),
  image loading will fail, regardless of whether its scheme is "https"
  or "javascript".
- 2. Otherwise, image should load successfully.

The test result should be:

|        | Flag off | Flag on |
|--------+----------+---------|
| Test 1 | PASS     | PASS    |
| Test 2 | FAIL     | PASS    |
| Test 3 | PASS     | PASS    |

In preparation for shipping the feature, we mark the test 2 as FAIL,
though all of the corresponding virtual/url-non-special tests are
marked as PASS.

When we ship the feature, all of the expectation will be rebased as
PASS.

[^1]: https://ci.chromium.org/ui/p/chromium/builders/try/linux-rel/1752691/test-results?sortby=&groupby=
[^2]: https://docs.google.com/spreadsheets/d/1Pqw1iKXK_lxHj-kLIAeRFs-khFz-BPZDio1W7SgEVE4/edit?usp=sharing

Bug: 40063064
Change-Id: Id7119d50e907961752e6b27119410c22a59f119b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5456854
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Commit-Queue: Hayato Ito <hayato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288603}
NOKEYCHECK=True
GitOrigin-RevId: 0346227cc62011c79c3a70c5e35e0243fb6785e2