Deflake attempt for fenced frame null config -> about:blank WPT
There are a few things wrong with this test (I was relatively new
to WPTs when I wrote it):
1. `attachFencedFrameContext` had no `await`, meaning that there
is no guarantee that the frame context exists before operating on it.
I'm assuming this is the reason for most of the failures, and have
fixed it.
2. The test had a probably-unnecessarily-long timeout, which I cut in
half to 1.5s
3. There was no meta timeout tag, which I've added just in case the
1.5s timeout causes breakage on slower builders.

I know that timeouts in tests are not ideal, but this test relies on
one because all other code that could be used to observe the fenced
frame is removed when navigating to about:blank.

Bug: 333501264
Change-Id: Ia54c69bde836f0b6bcb06a4cb1863d9de1d1c5e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5459098
Reviewed-by: Liam Brady <lbrady@google.com>
Commit-Queue: Andrew Verge <averge@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288725}
NOKEYCHECK=True
GitOrigin-RevId: 8977f42b8ca954f37ccb3df3baf049cf21b03aee