[css-color-adjust] Don't use opaque background for initial about:blank
When used color scheme for the embedder iframe and the root element of
the iframe document does not match we use an opaque background to avoid
mixing light/dark content through the transparent iframe. Do not do that
for the initial about:blank to avoid white/black flickering when the
iframe document takes a while to load.

Bug: 40947106
Change-Id: I80364df74098d52d71a50e93d4c00aaa52b62dac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5454400
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288845}
NOKEYCHECK=True
GitOrigin-RevId: 252444c8a4a65a65857fc873ec70fceeee63c947