Add wpt that asserts service worker involved navigation timing entry have correct sizes
We found that if a fetching is fulfilled by a service worker, the
corresponding navigation timing and resource timing would have 0
encodedBodySize and decodedBodySize.

This CL adds a wpt that asserts if an iframe is attached to a document
where a service worker is registered to intercept, the navigation timing
entry emitted on the subframe and the corresponding resource timing entry
emitted on the parent page should have correct encodedBodySize and
decodedBodySize.

Currently this wpt fails. We should fix the service worker to make it
pass.

This wpt is written by nrosenthal@ with additional asserts.

Bug: 334977808
Change-Id: I281abc24f85b92df6c23789fd1e810e2509c40bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5455553
Reviewed-by: Noam Rosenthal <nrosenthal@chromium.org>
Commit-Queue: Hao Liu <haoliuk@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288913}
NOKEYCHECK=True
GitOrigin-RevId: 2b386a50bbcf59e6d9008322fe6761a82969dd60