[PEPC] Add tests that check container CSS affects validation
There are various properties that can be applied to parent elements
that would affect the validity of the PEPC element. These are covered
by the IntersectionObserver visibility logic. This CL only adds tests
to verify the behavior.

Also a small bug fix to the "disable" logic where an indefinite disable
would not override a temporary disable. The scenario is verified in
*.IntersectionChangedDisableEnableDisable

Fixed: 335250088
Bug: 1462930
Change-Id: If0a6e1a7a3ba88de6ec43aea6e20944b81ebdd12
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5447702
Commit-Queue: Andy Paicu <andypaicu@chromium.org>
Reviewed-by: Thomas Nguyen <tungnh@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1289196}
NOKEYCHECK=True
GitOrigin-RevId: 23200bbcb12414babb68959c27d5893aca764e56