[fast-min-max] Avoid calling `HandleOverflow()` for min-content
This patch changes when `LineBreaker` is for `min-content`:
* Set the initial state to `kOverflow`. Since `min-content`
  computes the line breaking for 0 available width, the line
  almost always overflows.
* The `overflow-wrap` and `word-break` properties have fall
  back behaviors when the line overflows. This CL uses the
  fall back behavior from the first time.

It also copies the runtime flag to stack, as the stack access
is preferred over the global heap for the performance. Not
too much needed for this CL, but the following CL will need
it in per-character loops.

This patch should have no behavior changes as all the changes
are under a runtime flag.

Bug: 333630754
Change-Id: I541e60683cfcbf58c8468289d3f0167dbb8daec5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5465180
Commit-Queue: Koji Ishii <kojii@chromium.org>
Auto-Submit: Koji Ishii <kojii@chromium.org>
Commit-Queue: Kent Tamura <tkent@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1289202}
NOKEYCHECK=True
GitOrigin-RevId: aa39662aaf29e85aa3f92b3cc6694ad76c4d5053