[anchor] Call ResolveMathFunction on CSSValue after var() substitution
CL:5447517 incorrectly used the incoming CSSValue rather than
the result of ResolveSubstitutions. This caused a crash, because
we would not realize we were about to apply a CSSMathFunctionValue,
and therefore we would not discover invalid anchor() functions.

In addition to tests for var(), I added bonus tests to verify that
reverts do recurse into ResolveMathFunction when needed.
ResolveFlipRevert does actually not recurse into ResolveMathFunction,
but it's also not currently possible for ResolveFlipRevert to return
a CSSValue with different anchor()-validity than the unflipped value.
Added a test for this as well.

Note that we also have some theoretical bugs in the handling of
CSSAppearanceAutoBaseSelectValuePair, but I will address that
separately.

Fixed: 334377746
Change-Id: I076edfaad91f543e1ddd982aaf50f4102707c920
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5465622
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Daniil Sakhapov <sakhapov@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1289240}
NOKEYCHECK=True
GitOrigin-RevId: c29d8e1d8950af9f849b4545f5b4a013a1630f37