[css-scroll-snap-2] Set snapTargets to owning element of pseudo-element areas
Currently, JavaScript cannot interact with pseudo-elements in the same
way that it can with regular elements. This is raised in a spec
issue[1] where the current proposal is to treat snap events involving
pseudo-elements the same way that pointer events treat pseudo-elements:
let the snapTarget be the owning element.

[1] https://github.com/w3c/csswg-drafts/issues/10175

Bug: 40273052,40286359
Change-Id: I9e2a99de94eada4e06bcdab753d45f4ae46a2e46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5458351
Reviewed-by: Steve Kobes <skobes@chromium.org>
Commit-Queue: David Awogbemila <awogbemila@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1289306}
NOKEYCHECK=True
GitOrigin-RevId: 80220c17284dcea67a5aa6034416d48bc7896f66