[anchor] Resolve the result of TryValueFlips::FlipValue during cascade
We currently do not Resolve the result of TryValueFlips::FlipValue,
making the assumption that whatever TryValueFlips::FlipValue returns,
it does not need further resolution.

I *think* that happens to be true, but it's fragile: from the
perspective of the StyleCascade, we don't really know what FlipValue
will do exactly, and with its current implementation, we completely
rely on a certain AnchorEvaluator behavior, which is also outside
StyleCascade's control.

There should be no web-facing behavior change in this CL: it fixes
a theoretical problem that's only visible with the right setup
in unit tests.

Bug: 40059176
Change-Id: Ie63b7a6bece7bd695bd44d5819dbe264d219454a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5465624
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1289420}
NOKEYCHECK=True
GitOrigin-RevId: 5a070d0d01470b159262d45660fcaee2edd9221c