Hide deprecated prefixed fullscreen APIs
This CL downgrades the status of the deprecated PrefixedVideoFullscreen
runtime enabled feature. With this CL, the APIs will no longer be
available to websites, unless they register for the origin trial, which
is available starting in M125.

Enterprises can also set the PrefixedVideoFullscreenApiAvailability
policy to temporarily re-enable the deprecated APIs, if an OT is not
feasible in their internal websites.

Link to blink-dev discussion concerning this deprecation:
https://groups.google.com/a/chromium.org/g/blink-dev/c/jrkCYeUSusg/m/1n7mUe89AQAJ

Bug: 40352864
Change-Id: Ia1d776ef9e9334a8fc96032ec7155ccd40dd21a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5463779
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1289456}
NOKEYCHECK=True
GitOrigin-RevId: c9fb12ce3adc0f056e06a358bc297e548e886383