compute pressure: Cancel pending tasks on Unobserve
When using rate obfuscation mitigation, if a penalty is applied, the
last record is saved and will be send to the user after the penalty.
A task is posted using PostDelayedCancellableTask.

If the user calls `unobserve` for a designated `source` or
calls `disconnect`, all the records are cleaned, but the
delayed task is not cancelled.

This patch fixes this issue by also cancelling the delayed task
on `unobserve` and `disconnect` calls.

Bug: 333062551
Change-Id: Ia63ba0cc644d4b78a121ee76dfd3f62b2234b831
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5434975
Reviewed-by: Raphael Kubo Da Costa <raphael.kubo.da.costa@intel.com>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Arnaud Mandy <arnaud.mandy@intel.com>
Cr-Commit-Position: refs/heads/main@{#1289794}
NOKEYCHECK=True
GitOrigin-RevId: 15a5e3482b7972dc0a99ab9aa2b2d400b5d7d1d3