Clean up expired flag 'add-identity-in-can-make-payment'
This flag expired back in M115. It was part of an opt-out origin
trial to restore certain features to the CanMakePaymentEvent IDL
file. Now that it has expired, those features can be removed entirely.

We may re-introduce a similar flag in future if/when we tackle
native app IS_READY_TO_PAY, but there is no point in keeping this
code/flag around for that potential future.

Bug: 40909272, 40212375

Change-Id: I31aa01b9406685b88de16cec5a789d21f53b2815
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5457560
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1290028}
NOKEYCHECK=True
GitOrigin-RevId: d28168fbdcf0c1898cb01be77fcac653b8a3a23b