Add helper class to manage JavaScript call stack collection.
There is an issue where `Check failed:
connected.CallStackGenerator::CollectJavaScriptCallStackCallback
was destroyed without first either being run or its corresponding
binding being closed. It is an error to drop response callbacks
which still correspond to an open interface pipe.`

I suspect this to be because the
`CollectJavaScriptCallStackCallback` stored in `callback_` was
being overwritten due to consecutive calls.

The solution involves assigning each Mojo callback to an
instance of a helper class(JavaScriptCallStackCollector) for
each call to `CollectJavaScriptCallStack()`. This class is
responsible for collecting the call stack and posting a task
to execute the Mojo callback.

a hashmap is used to manage the lifetime of these helper class
instances, ensuring they are kept alive until their work is
complete and then properly destroyed. This approach appears to
have resolved the flaky test issue based on CI test results.

An alternative solution using
`mojo::WrapCallbackWithDefaultInvokeIfNotRun` was attempted
but did not resolve the issue.

Bug: 330909120,40268201
Change-Id: Ifbcdd216feebce5df3ca7247d5ccb8e432590570
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5459362
Reviewed-by: Seth Brenith <seth.brenith@microsoft.com>
Commit-Queue: Issack John <issackjohn@microsoft.com>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1290265}
NOKEYCHECK=True
GitOrigin-RevId: 7cc4c65ac1a327b16697a3deb1f65452b50cec56