Apply fast rounded corner in correct transform space
Previously the fast border radius was applied in the transform space
of the mix-blend-mode, causing unexpected offset. We had a logic to
avoid fast border radius in the case for backdrop-filter, and could
avoid the problem if the logic also applied to mix-blend-mode.

Now adjust the rounded rect if the transform difference is a 2d
translation. This applies to all backdrop effects including
mix-blend-mode and backdrop-filter. This not only fixes the bug for
mix-blend-mode, but also enables fast rounded corner for
backdrop-filter when possible.

Fixed: 328339028
Change-Id: I2577ce36f2d5d9ba89a1dcd57fb3d5fc2b240485
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5463199
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Auto-Submit: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1290277}
NOKEYCHECK=True
GitOrigin-RevId: a1c0225b18d4ee214aa6ce36a1daac072a4101df