Rearrange fuzzing engine dependencies
We are working towards allowing fuzztests in arbitrary existing unit
test suites. This is a step in that direction.

Specifically, base/test/test_suite.cc will need to initialize FuzzTest,
and that means it will come to depend upon fuzztest.

Previously, this would have meant it transitively came to depend upon
some of the targets within //testing/libfuzzer which are used to denote
actual fuzz targets. These marker targets are used by recipes to
determine which fuzz targets to build, and thus these builders would
have started to build all the various unit test targets.

Not only is this slow, but some of those targets don't build OK in the
builder configurations we use for fuzzers.

This CL rearranges things as follows:

* It introduces a new fuzzing_engine_no_main_core target, which
  test_support can (transitively) depend upon, but does not indicate
  that the target is a fuzzer which needs building.

(this is in fact the only change needed to solve the bug as described,
but a few other tidy ups in the area to reduce technical debt as
follows):

* The existing 'fuzzing_engine' target is renamed to 'is_a_fuzz_target'
  to make its unusual purpose as a marker group clearer.
* The various fuzztest test suites themselves no longer depend upon
  'is_a_fuzz_target'. Instead, the small wrapper targets which trigger
  individual fuzz tests depend upon 'is_a_fuzz_target' and they
  trigger the building of the main fuzztest by means of a data_deps
  dependency.
* The centipede driver executable no longer depends upon
  'is_a_fuzz_target' (since it isn't a fuzz target). Instead it gets
  build by means of a data_deps relationship.

Bug: 40286211
Cq-Include-Trybots: luci.chromium.try:linux-libfuzzer-asan-rel,win-libfuzzer-asan-rel,linux-centipede-asan-rel
Change-Id: I25e7cdc67fc15246af13d22d91feb60c9ce71886
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5465828
Commit-Queue: Adrian Taylor <adetaylor@chromium.org>
Reviewed-by: Paul Semel <paulsemel@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1290623}
NOKEYCHECK=True
GitOrigin-RevId: a08510c1fcd0d87ee0f08b09fe707c7d7b3d5bf4