[PEPC] Provide partial padding support to allow for common use case
Removal of 'padding' support from the permission element has removed
support for a common use case that used padding in conjunction with
'width: auto'. Since it is very difficult to provide a fixed width
for the permission element (as the text can change and is not under
the control of the site author), this CL provides a mechanism that
sets the width/height based on padding.

We don't want to simply support padding unchecked since it can be
used to push the text outside of the permission element area, and
creating a bounded expression is very difficult because 'calc-size'
is not supported for padding lengths.

Also do a quick cleanup of vertical-align.

Bug: 1462930
Fixed: 335835011
Change-Id: I19e6f9ff6fcd11073b1da0efaa68f743ea3886bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5458152
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Andy Paicu <andypaicu@chromium.org>
Reviewed-by: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1291274}
NOKEYCHECK=True
GitOrigin-RevId: c03a7a5f518a45365543197744c349ccfe5ce613