[Device Provider] Inject PresenceDeviceProvider in NC
Inject a pointer to the PresenceDeviceProvider into NearbyConnections on
initialization. The DeviceProvider will not be registered until
RegisterServiceWithPresenceDeviceProvider is called and will only be
registered for the service specified by the service_id.

This cl depends on third_party changes reviewed in: cl/625447691

Sharing is still working

Test: Added unittests and manually tested on DUT to confirm Nearby
Change-Id: Ic559909a5675cea008e7c0d3b8dfe53f964f072f
Bug: b/330183112
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5458144
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Reviewed-by: Chris Bookholt <bookholt@chromium.org>
Commit-Queue: Ryan Hansberry <hansberry@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1291341}
NOKEYCHECK=True
GitOrigin-RevId: 6073fe759474359058d4b07109c07681f1663b40