Disable anti-aliasing for the Ahem font
The current anti-alias behavior of Ahem is responsible for many WPT
test failures on Windows. The fix for this is to opt Ahem out of anti
aliasing. This is safe, as Ahem is a test-only font with hard edges,
so anti-aliasing is just adding unnecessary pixels.

Both Safari and Firefox have similar behavior, as noted in the WPT
issue - https://github.com/web-platform-tests/wpt/issues/45750.

This behavior is gated behind the new (off-by-default) Runtime Enabled
Feature `DisableAhemAntialias`.

Bug: 335269878
Change-Id: Icec6b0577a11c9b8ea111b10a469052e86052521
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5458787
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1291529}
NOKEYCHECK=True
GitOrigin-RevId: 71fcfe4ba7bf7cdcb1dc185a8461a584b9bd4634