[FedCM] Record a use counter when we get a continue_on response
This is a requirement for origin trials.

Only a subset of FedCM requests will use this feature, and we expect
this to be a very small subset. FedCM as a whole is at 0.6%
(https://chromestatus.com/metrics/feature/timeline/popularity/4166)
and this should be a small percentage of that, I expect at most
a single-digit percentage of those calls, more likely a lot less
(and even that assumes a full rollout of our partners)

Bug: 324104289
Change-Id: Id84f010b55fdd56f48b5fe80d89c68976609225f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5474124
Auto-Submit: Christian Biesinger <cbiesinger@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Christian Biesinger <cbiesinger@chromium.org>
Reviewed-by: Yi Gu <yigu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1291568}
NOKEYCHECK=True
GitOrigin-RevId: d71d314c49eb806a5d0a509d6575613805ea9c62