Fix <select> native popup with <datalist> <option>s
In http://crrev.com/1285415 I made the following markup successfully
render options in the native popup in appearance:auto mode:
<select>
  <datalist>
    <option>option</option>

However, it didn't completely work, at least on MacOS (not sure how
other platforms worked), because the <option>s won't have a computed
style because the <datalist> is display:none even when the native popup
is open. display:none options are not supposed to be included in the
native popup as a way to allow developers to exclude options from being
included in the native popup.

This patch fixes this by adding a select:open rule to make the
<datalist> display:block so that the <option>s can have a computed
style.

Bug: 1511354
Fixed: 334122678
Change-Id: I2ef80ed25feacef9f62cf7cee68347110081be4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5462883
Reviewed-by: David Baron <dbaron@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1291607}
NOKEYCHECK=True
GitOrigin-RevId: 0fe7e33702cb13713a0474b95459516cb44073ec