Simplify logic in |Range::BoundaryNodeChildrenWillBeRemoved|.
Current logic can sometimes have perf issues when the to-be removed
container node has too many children because
|Range::BoundaryNodeChildrenWillBeRemoved| iterates through all the
children of the container node and check every cached range if the
child of container node is an ancestor container of the range. This
patch is to simplify the logic and improve performance by adjusting
the boundary points of the range to the start of the container node if
the container node contains the range. The change should have no
behavior change.

The perf improvement is huge when there are lots of children (~50k) in
container node. For each BoundaryNodeChildrenWillBeRemoved call: 3.7ms
vs 0.003 ms (1000x faster).

|Range::BoundaryShadowNodeChildrenWillBeRemoved| can also be simplified.
Will submit a separate CL to address the
issue(https://issues.chromium.org/issues/336222283).

Bug: 328954155
Change-Id: I391c9be138694fe27b02d1bfac1b7761523c7d19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5414145
Reviewed-by: Sanket Joshi <sajos@microsoft.com>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Siye Liu <siliu@microsoft.com>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1291945}
NOKEYCHECK=True
GitOrigin-RevId: cf4f730829d9bb939192d777c04bbabd5042d933