Rename Length::IsPercentOrCalc to MayHavePercentDependence.
Following https://chromium-review.googlesource.com/c/5410698 the
remaining callers of Length::IsPercentOrCalc are all callers that are
making an optimization to check whether they can optimize away a
computation of the percentage basis or they need to handle a change to
the percentage basis.

This renames Length::IsPercentOrCalc to Length::MayHavePercentDependence
so that callers who might have previously used IsPercentOrCalc are
pushed to HasPercent when it is more appropriate, and adds additional
documentation for the method.

This converts the callers that are (or may be for) for properties that
can contain calc-size() to use MayHavePercentDependence(), and converts
the callers that are for properties that cannot (and where the
distinction doesn't matter) to use the shorter HasPercent().

See thread at
https://groups.google.com/a/chromium.org/g/layout-dev/c/ksyQUk08nqs/m/VyEyVDc9AAAJ

Bug: 313072
Change-Id: I1ba3992d2bee81c67997e25970d18d92f332547e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5427589
Commit-Queue: David Baron <dbaron@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1292024}
NOKEYCHECK=True
GitOrigin-RevId: 08d9c71ac53457e12fa84e0985944179df9c0a72