WPT: COEP: require-corp should also be set in the loaded script
Since the WPT uses shared array buffer, it should have a COEP header.
The HTML files of the WPTs have both, but the worker scripts do not.
It should bring COEP violation when the script is loaded.
This CL adds proper COEP headers to fix the violation.

Note that even with this change, Chromium still fails with timeout
because code to create a dedicated worker is not fully parallel.
The worker creation is currently blocked by busy-looping JavaScript.
(See crbug.com/40687798).

Bug: 334026778, 40687798
Change-Id: I86ad8dde5b6d21aef88ab18e63a81f9486b3e276
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5480214
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1292311}
NOKEYCHECK=True
GitOrigin-RevId: ddd32f326fbd909600b44cb96a501f72c9195138