Use IsolatedSVGDocumentHost in SVGResourceDocumentContent
Replace SVGResourceDocumentContent's Document with a
IsolatedSVGDocumentHost, and make the required adjustments to handle
async completion. Add IsolatedSVGChromeClient as a superclass for
SVGImageChromeClient, and then add an implementation for the
SVGResourceDocumentContent embedding.

The cache class is renamed to SVGResourceDocumentCache and made public.
It's then moved to be hosted by the Page so that disposal/shutdown of
the isolated document can be performed in a controlled manner. This is
performed outside of the "regular" GC, since shutting down the inner
Page is a bit too complex of an operation.

Bug: 40134477
Change-Id: I99023995077ff3c1c2dd74ca9dcadc4415e83a76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5458161
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1292323}
NOKEYCHECK=True
GitOrigin-RevId: 4f3ebd119d5c8492fa8564d61d8b600a82e4a041