Handle missing optional step (B) argument in CSS round() function
The following part of the spec[1] was not handled properly:

  If the type of A matches <number>, then B may be omitted, and defaults
  to 1; omitting B is otherwise invalid.

Canonicalize the parameter list by adding B=1 when the above is
fulfilled. Update the serialization code to prefer the shortest form by
dropping B if it can be omitted per above.

[1] https://drafts.csswg.org/css-values-4/#round-func

Bug: 336681966
Change-Id: I582c11e352b3b26f91c840fc97ac16b6787fdcc4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5487527
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Daniil Sakhapov <sakhapov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1292428}
NOKEYCHECK=True
GitOrigin-RevId: 25ae577fbd442e8872592b3817ddd749ed27520e