More dumps to investigate AnchorElementMetricsSender::UpdateMetrics
The checks related to metrics_partitions_ hold, so we turn those into
full checks.

We continue to investigate why the EraseIf assertions don't hold.
A possibility could be that an anchor gets an id that isn't valid to
store in a map. We dump if that happens.

Bug: 331043758
Change-Id: I604ec08893897713fac96f4e734c6d793f36bef8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5480667
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1292597}
NOKEYCHECK=True
GitOrigin-RevId: 65caa4cacc1ef93126a0f8b7bbdbc5c4cca71b97