Replace dead link documenting HashInt with an archived version
The link is dead. It looks like it was last available in 2012, but the
content of the page at that point differs from what's implemented in
blink. A version of the page from 2006 has the functions, so we link to
that one.

Bug: None
Change-Id: Ibca043ec7afb918c4807fcf26df4b9beb342315e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5485927
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1292598}
NOKEYCHECK=True
GitOrigin-RevId: 48004632bca407223ad3ce7d709011a8ae3f81b2