Add a RuntimeEnabledFeature for FedCM/SAA autogrants
This CL prepares for an origin trial. It swaps out the handwritten,
disabled-by-default base::Feature for a REF-generated,
enabled-by-default Feature as well as a RuntimeEnabledFeature. This is
necessary for the OT because the implementation relies on some
out-of-band syncing that cannot be triggered by the pages loaded on
OT origins; the syncing occurs in the background and therefore must
occur by default, even if most pageloads run with the REF disabled.

The generated Feature can be enabled/disabled by the user or by field
trial, and the overridden state takes precedence over the origin trial.

Bug: 330574529
Change-Id: Ia6749fe15a9ecdc864292d7a6362b8888ff3901d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5472651
Commit-Queue: Chris Fredrickson <cfredric@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293024}
NOKEYCHECK=True
GitOrigin-RevId: d459f7b34291c4cff9be5411dfd9311312938f68