Stop using TimestampAligner in WebrtcAudioSink
As WebRTC clocks and Chrome clocks are the same, there is no reason to
use a TimestampAligner.

This change also add a feature so it is possible to get back the
timestamp aligner if one notice it is causing issue.

Bug: 328203941
Change-Id: If2f1e6480e66464687e01d21e8dc1b18857e6af7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5482078
Reviewed-by: Markus Handell <handellm@google.com>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Lionel Koenig Gélas <lionelk@google.com>
Cr-Commit-Position: refs/heads/main@{#1293036}
NOKEYCHECK=True
GitOrigin-RevId: ba2e694e9934202882044018739963eb1ccd6a26