Get rid of some of V8DOMWrapper::IsDOMWrapper() calls
All of these can be replaced with concrete HasInstance() calls.
The remaining ones are a little bit more complicated.

As a drive-by, also simplify context snapshotter with assumption
that the only kind of wrapper that actually exists there is one for
the document.

Bug: 328117814
Change-Id: I163f502a6c7cb7081da7aea1b5a816bb0fffc1d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5474123
Reviewed-by: Nate Chapin <japhet@chromium.org>
Commit-Queue: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293269}
NOKEYCHECK=True
GitOrigin-RevId: 1e7c11bf6f4ab832970aad00f42f6fab0a466e2b