[blinkpy] Include stderr in `ScriptError` of failed `git cl` commands
`git cl try` will fail if the CL is closed, but that information is
logged to stderr [0], which `GitCL` currently redirects to DEVNULL.

Redirect this information to PIPE instead, which is only retrieved when
`Executive.run_command(...)` raises a `ScriptError`. This ensures
`GitCL.run(...)` still returns only stdout when successful, so as not to
interfere with callers' parsing.

[0]: https://chromium.googlesource.com/chromium/tools/depot_tools/+/85e409e6/git_cl.py#5702

Fixed: 336964853
Test: git cl issue 5488337 && ./blink_tool.py rebaseline-cl
Test: "CL is closed" is logged
Change-Id: Ia67e40796ef7a0184178ad00403b546b80f5c23a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5494697
Commit-Queue: Jonathan Lee <jonathanjlee@google.com>
Reviewed-by: Weizhong Xia <weizhong@google.com>
Cr-Commit-Position: refs/heads/main@{#1293350}
NOKEYCHECK=True
GitOrigin-RevId: af3f06810f71c09846ed26a6b20394607501ffa3